+++ platform/processor/live-view/src/main/kotlin/spp/processor/view/impl/LiveViewServiceImpl.kt
@@ -59,7 +59,7 @@ import spp.processor.view.impl.view.LiveMeterView
 import spp.processor.view.impl.view.LiveTraceView
 import spp.processor.view.impl.view.util.EntitySubscribersCache
 import spp.processor.view.impl.view.util.MetricTypeSubscriptionCache
-import spp.processor.view.impl.view.util.ViewSubscriber
+import spp.processor.view.model.ViewSubscriber
 import spp.processor.view.model.LiveMeterConfig
 import spp.processor.view.model.LiveMetricConvert
 import spp.processor.view.model.LiveMetricConvert.Companion.NOP_RULE
