+++ platform/processor/live-view/src/main/kotlin/spp/processor/view/model/ViewSubscriber.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package spp.processor.view.impl.view.util
+package spp.processor.view.model
 
 import io.vertx.core.eventbus.MessageConsumer
 import io.vertx.core.json.JsonObject
